Move code around to better match e2e build
